refactor: constructor 구현 lombok annotation 으로 대체 (리뷰반영)
