+++ src/main/java/nextstep/subway/ui/StationController.java
@@ -1,6 +1,7 @@
 package nextstep.subway.ui;
 
-import nextstep.subway.applicaion.StationService;
+import lombok.RequiredArgsConstructor;
+import nextstep.subway.applicaion.station.StationService;
 import nextstep.subway.applicaion.dto.StationRequest;
 import nextstep.subway.applicaion.dto.StationResponse;
 import org.springframework.http.MediaType;
@@ -11,12 +12,10 @@ import java.net.URI;
 import java.util.List;
 
 @RestController
+@RequiredArgsConstructor
 public class StationController {
-    private StationService stationService;
 
-    public StationController(StationService stationService) {
-        this.stationService = stationService;
-    }
+    private final StationService stationService;
 
     @PostMapping("/stations")
     public ResponseEntity<StationResponse> createStation(@RequestBody StationRequest stationRequest) {
